backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
bitarray==2.7.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
cytoolz==0.12.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
hexbytes==0.3.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.2
multidict==6.0.4
orjson==3.8.9
parsimonious==0.9.0
passlib==1.7.4
pip==21.3.1
protobuf==4.22.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pymongo==4.3.3
pyrsistent==0.19.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rlp==3.0.0
rsa==4.9
setuptools==60.2.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
toolz==0.12.0
typing-extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
web3==6.0.0
websockets==10.4
wheel==0.37.1
yarl==1.8.2
